Server/s_requirements.txt
bcrypt
pymongo
pycryptodome
python-dotenv
groq
langchain-groq